Refactor backgroundscript external messages handling
